sync: main → develop 동기화 (prop 타입 단일원천화 D+B 포함) (#124)
* chore: Deploy DatePOP Claude Code marketplace settings

* chore(release): bump @pop-ui/{foundation,core}@1.1.9

PR #107 (foundation 아이콘·일러스트 추가 + brand 카테고리 신설)를 NPM으로 publish
하기 위한 monorepo 단일 버전 bump. v1.1.9 태그 push 시 Jenkinsfile의 release
validation을 통과시키기 위해 foundation/core 양쪽을 동일하게 올림.


* chore(release): bump @pop-ui/{foundation,core}@1.1.10


* [DPN-1654] develop 변경사항을 release에 동기화 (#119)

* chore: Deploy DatePOP Claude Code marketplace settings

* chore(release): bump @pop-ui/{foundation,core}@1.1.9

PR #107 (foundation 아이콘·일러스트 추가 + brand 카테고리 신설)를 NPM으로 publish
하기 위한 monorepo 단일 버전 bump. v1.1.9 태그 push 시 Jenkinsfile의 release
validation을 통과시키기 위해 foundation/core 양쪽을 동일하게 올림.


* chore(release): bump @pop-ui/{foundation,core}@1.1.10


---------


* refactor(foundation): export border-radius tokens as JS constants; use BorderRadius150 in SegmentButton


* refactor(core): rename type.ts to types.ts in Button and Toast for consistency


* refactor(core): restore I/T type prefixes in ImageInput and LottieInput


* refactor(core): unify component exports to named-only


* docs: add CLAUDE.md with pop-ui API conventions for AI/human consistency


* chore(foundation): fix stale typography token name in variables.scss

web-h1-36 → web-h1-40 이름 정합(값 동일, font-size는 이미 40이었음). 구버전 style-dictionary로 생성됐던 stale 산출물 정리.

colors.ts는 style-dictionary(큰따옴표) vs prettier(작은따옴표) 차이였을 뿐 값 변화 없음 — prettier가 관리하므로 이미 정상 상태라 제외.


* chore: stop prettier from reformatting generated token files

style-dictionary는 큰따옴표로 생성하는데 prettier(singleQuote)가 매번 작은따옴표로 되돌려 token-build마다 무의미한 diff 발생. generated tokens를 .prettierignore에 추가해 drift 소멸.


* style(foundation): fix import group spacing in Illustrations.stories

기존 lint 에러 2건(import/order) 해소. eslint --fix 결과.


* fix(button): align loading behavior with spec — hide label, intercept Mantine loading

- hideLabelOnLoading 기본값 false→true: 로딩 중 기본으로 라벨 숨기고 스피너만 노출
- Mantine loading prop을 로딩 트리거로 수용하되 구조분해로 가로채 unstyled 모드에서
  깨지는 Mantine 로더로 전달되지 않게 함(isLoading || loading)
- 기존 실패하던 Button 테스트 23개 통과 (총 33 pass)


* fix(core): re-export Button as named instead of default

Task 4에서 Button/index.tsx의 default export를 제거했으나 components/index.ts의
재수출이 여전히 `export { default as Button }`로 남아있던 누락 수정.
다른 컴포넌트와 동일하게 named 재수출로 통일.


* refactor(core): extract inline prop types to types.ts (simple cases)

Checkbox, Dropdown, Modal, Pagination, Radio, SearchBar, SegmentButton,
Tab, TimePicker, Tooltip — move inline IXxxProps into dedicated types.ts.
No runtime change.


* refactor(core): extract IToggleProps to types.ts, use named ChangeEvent


* refactor(core): extract IDatePickerProps to types.ts, use named ReactNode

Standalone interface with external type imports (TDateDisplayType,
TDayOfWeek, PopoverProps). No runtime change.


* refactor(core): extract TextField prop types to types.ts

Move ICommonTextFieldProps helper + TTextFieldProps union into types.ts.
Use named ChangeEvent. ICommonTextFieldProps stays exported (body uses it
in type casts). No runtime change.


* feat(core): export all component prop types from barrel

Expose prop types (IButtonProps 등 15개 컴포넌트) as public API. Values
(BUTTON_SIZES 등) exported separately from types. No existing export
removed/renamed — not breaking (minor).


* docs: 컨벤션에 prop 타입 types.ts 정의 + barrel export 규칙 추가

- prop 타입은 types.ts에 정의(인라인 금지), barrel에서 export
- union prop 타입은 TXxxProps 유지(eslint가 type alias에 T 강제),
  IXxxProps 별칭 안 만듦


---------
