+++ packages/core/src/components/DatePicker/index.tsx
@@ -18,14 +18,13 @@ import {
   resolveDatePickerValue,
 } from '../CalendarDatePicker/utils';
 
+import type { IDatePickerProps } from './types';
 import type {
   TMantineClassNameRecord,
   TMantineClassNamesResolver,
-  TDayOfWeek,
 } from '../CalendarDatePicker/types';
 import type { PopoverProps, PopoverStylesNames } from '@mantine/core';
 import type { DateTimePickerProps, DateTimePickerStylesNames, DateValue } from '@mantine/dates';
-import type { TDateDisplayType } from '@pop-ui/foundation';
 
 const DEFAULT_DATE_TIME_CLASS_NAMES = {
   wrapper: styles.DatePicker__Wrapper,
@@ -83,37 +82,6 @@ const toDateOrNull = (v: DateValue | null | undefined): Date | null => {
   }
   return null;
 };
-
-export interface IDatePickerProps {
-  size?: 'sm' | 'md' | 'lg';
-  type?: TDateDisplayType;
-  withTime?: boolean;
-  // 값 제어
-  value?: string | null | string[] | [string | null, string | null];
-  defaultValue?: string | null | string[] | [string | null, string | null];
-  onChange?: (value: string | null | string[] | [string | null, string | null]) => void;
-  // Input 외관
-  placeholder?: string;
-  label?: React.ReactNode;
-  description?: React.ReactNode;
-  error?: React.ReactNode;
-  disabled?: boolean;
-  clearable?: boolean;
-  // 날짜 제약
-  minDate?: Date;
-  maxDate?: Date;
-  displayValueFormat?: string; // 입력창 표시 형식. 기본: 'YYYY년 MM월 DD일'
-  valueFormat?: string; // onChange emit 형식. 기본: 'YYYY-MM-DD'
-  // CalendarDatePicker 전용
-  excludedDates?: (string | [string, string])[];
-  excludedDays?: TDayOfWeek[];
-  highlightToday?: boolean;
-  // 스타일
-  className?: string;
-  popoverProps?: PopoverProps;
-  rightSection?: React.ReactNode;
-  rightSectionWidth?: number;
-}
 
 export const DatePicker = ({
   size = 'md',
