+++ packages/core/src/components/Pagination/types.ts
@@ -1,0 +1,9 @@
+import type { HTMLAttributes } from 'react';
+
+export interface IPaginationProps extends HTMLAttributes<HTMLDivElement> {
+  currentPageIdx: number;
+  rowsPerPage?: number;
+  totalLength?: number;
+  paginationSize?: number;
+  onPageChange?: (e: number) => void;
+}
