+++ packages/core/src/components/SearchBar/index.tsx
@@ -6,19 +6,7 @@ import { useCallback, useState } from 'react';
 
 import styles from './styles.module.scss';
 
-import type { AutocompleteProps } from '@mantine/core';
-
-export interface ISearchBarProps extends AutocompleteProps {
-  label?: string;
-  labelPosition?: 'top' | 'left';
-  size?: 'sm' | 'md' | 'lg';
-  tooltip?: string;
-  tooltipPosition?: 'top' | 'bottom' | 'left' | 'right';
-  description?: string;
-  errorMsg?: string;
-  onChange?: (value: string) => void;
-  onClear?: () => void;
-}
+import type { ISearchBarProps } from './types';
 
 export const SearchBar = ({
   label,
