+++ packages/core/src/components/SegmentButton/types.ts
@@ -1,0 +1,6 @@
+import type { SegmentedControlProps } from '@mantine/core';
+
+export interface ISegmentButtonProps extends SegmentedControlProps {
+  size?: 'sm' | 'md' | 'lg';
+  radius?: number | 'xs' | 'sm' | 'md' | 'lg' | 'xl';
+}
