+++ packages/core/src/components/Tab/index.tsx
@@ -4,18 +4,7 @@ import { Tabs } from '@mantine/core';
 
 import styles from './styles.module.scss';
 
-import type { TabsProps } from '@mantine/core';
-import type { ReactNode } from 'react';
-
-export interface ITabProps extends TabsProps {
-  tabList: {
-    title: string;
-    value: string;
-    body: ReactNode;
-    icon?: ReactNode;
-  }[];
-  containerPaddingTop?: 'xs' | 'sm' | 'md' | 'lg' | 'xl';
-}
+import type { ITabProps } from './types';
 
 export const Tab = ({ tabList = [], containerPaddingTop, ...props }: ITabProps) => {
   return (
