+++ packages/core/src/components/TimePicker/types.ts
@@ -1,0 +1,5 @@
+import type { TimeInputProps } from '@mantine/dates';
+
+export interface ITimePickerProps extends TimeInputProps {
+  size?: 'sm' | 'md' | 'lg';
+}
