+++ packages/core/src/components/index.ts
@@ -1,9 +1,14 @@
 export { Button } from './Button';
+export type { IButtonProps, TButtonSize, TButtonVariant } from './Button/types';
+export { BUTTON_SIZES, BUTTON_VARIANTS } from './Button/types';
 export { Checkbox } from './Checkbox';
+export type { ICheckboxProps } from './Checkbox/types';
 export { CalendarDatePicker } from './CalendarDatePicker';
 export type { ICalendarDatePickerProps, TDayOfWeek } from './CalendarDatePicker';
 export { DatePicker } from './DatePicker';
+export type { IDatePickerProps } from './DatePicker/types';
 export { Dropdown } from './Dropdown';
+export type { IDropdownProps } from './Dropdown/types';
 export { ImageInput } from './ImageInput';
 export type {
   TImageInputAccept,
@@ -20,16 +25,27 @@ export type {
   ILottieInputProps,
 } from './LottieInput/types';
 export { Modal } from './Modal';
+export type { IModalProps } from './Modal/types';
 export { Pagination } from './Pagination';
+export type { IPaginationProps } from './Pagination/types';
 export { Radio } from './Radio';
+export type { IRadioProps } from './Radio/types';
 export { SearchBar } from './SearchBar';
+export type { ISearchBarProps } from './SearchBar/types';
 export { SegmentButton } from './SegmentButton';
+export type { ISegmentButtonProps } from './SegmentButton/types';
 export { Tab } from './Tab';
+export type { ITabProps } from './Tab/types';
 export { TextField } from './TextField';
+export type { TTextFieldProps } from './TextField/types';
 export { TimePicker } from './TimePicker';
+export type { ITimePickerProps } from './TimePicker/types';
 export { toast } from './Toast';
+export type { IToastOptions, TToastInput } from './Toast/types';
 export { Toggle } from './Toggle';
+export type { IToggleProps } from './Toggle/types';
 export { Tooltip } from './Tooltip';
+export type { ITooltipProps } from './Tooltip/types';
 
 // Map components
 export {
